End of Day 22 Jan
DataController and Level Controller refactoring and events added.

cockpit 3d assets added.

+++ Assets/Scripts/GameData.cs
@@ -3,20 +3,20 @@ using System.Collections.Generic;
 using UnityEngine;
 
 /// <summary>
-/// Stores active Game Data for Json Operability.
+/// Stores active Game Data for Json Interoperability.
 /// </summary>
 public class GameData 
 {
     //This only needs to be updated when I am prepping save and load from file operations.
 
     //TODO: Ensure this is synced with active play classes: Player, Ship, Pilot
-    //TODO: Method: UpdateSaveGameDataWithPlayerData() & UpdatePlayerDataWithSavedGameData() in datacontroller
+    //TODO: Method: UpdateSaveGameDataWithPlayerData() & UpdatePlayerDataWithSavedGameData() in DataController
     //need to be updated with any property / variable changes
 
 
 
 
-    public float selectedPilotIndex;
+    public int selectedPilotIndex;
